Version 4.0 Released
This version adds the much anticipated 'Color Image Mapping' feature.
With this feature, it is now possible to generate mazes in custom areas
as defined by cell color in the ImageMapping sheet. Included with this
commit is a file which takes a BMP file and applies it to cells. An
example of the output using this feature has been uploaded to
demonstrate how cool it is.

- Fix color bug, color now adjusts properly during spill
- Previous spill function was nice too though, now titled Insane Spill
under Special Options
- Image Mapping image included for demoing